Add BeePx
I am not sure if it fits the "Clones" category, but not to me to decide, therefore I propose adding my own TypeScript framework BeetPx.(https://beetpx.dev/), which focuses on a web as a target platform. It's heavily inspired by PICO-8 (I ported some of my PICO-8 games to it), e.g. it's also based on a fixed time step loop, have a square canvas ratio, and the drawing API is kinda a clone in its shape.